Módulo 4 - Rest com Spring

+++ algafood-api/src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,8 +26,9 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 	
 	@Transactional 
 	@Override
-	public void remover(Cozinha cozinha) {
-		cozinha = this.buscar(cozinha.getId());
+	public void remover(Long id) {
+		Cozinha cozinha = this.buscar(id);
+		if (cozinha == null) throw new EmptyResultDataAccessException(1);
 		entityManager.remove(cozinha);
 	}
